🎨 [clippy]

+++ src/input.rs
@@ -119,7 +119,7 @@ impl SystemState {
 
     pub fn aspect_ratio(&self) -> f32 {
         let width = self.window_dimensions.width;
-        let height = std::cmp::max(self.window_dimensions.height as u32, 0);
+        let height = std::cmp::max(self.window_dimensions.height, 0);
         width as f32 / height as f32
     }
